requirements.txt
fastapi == 0.112.2
uvicorn == 0.30.6
firebase-admin == 6.5.0
google-cloud-firestore == 2.18.0
discord.py==2.4.0